create principal from der-encoded public key
