+++ ic-agent/src/identity/basic.rs
@@ -1,7 +1,10 @@
 use crate::export::Principal;
 use crate::{Identity, Signature};
+use num_bigint::BigUint;
 use ring::signature::{Ed25519KeyPair, KeyPair};
 use thiserror::Error;
+use simple_asn1::{OID, to_der, ASN1EncodeErr};
+use simple_asn1::ASN1Block::{Sequence, ObjectIdentifier, BitString};
 
 /// An error happened while reading a PEM file to create a BasicIdentity.
 #[derive(Error, Debug)]
@@ -20,6 +23,7 @@ pub enum PemError {
 /// A Basic Identity which sign using an ED25519 key pair.
 pub struct BasicIdentity {
     key_pair: Ed25519KeyPair,
+    der_encoded_public_key: Vec<u8>
 }
 
 impl BasicIdentity {
@@ -36,20 +40,21 @@ impl BasicIdentity {
             .bytes()
             .collect::<Result<Vec<u8>, std::io::Error>>()?;
 
-        Ok(Self {
-            key_pair: Ed25519KeyPair::from_pkcs8(pem::parse(&bytes)?.contents.as_slice())?,
-        })
+        Ok(BasicIdentity::from_key_pair(Ed25519KeyPair::from_pkcs8(pem::parse(&bytes)?.contents.as_slice())?))
     }
 
     /// Create a BasicIdentity from a KeyPair from the ring crate.
     pub fn from_key_pair(key_pair: Ed25519KeyPair) -> Self {
-        Self { key_pair }
+        let der_encoded_public_key = der_encode_public_key(key_pair.public_key().as_ref().to_vec())
+            .expect("DER encoding error");
+
+        Self { key_pair, der_encoded_public_key }
     }
 }
 
 impl Identity for BasicIdentity {
     fn sender(&self) -> Result<Principal, String> {
-        Ok(Principal::self_authenticating(&self.key_pair.public_key()))
+        Ok(Principal::self_authenticating(&self.der_encoded_public_key))
     }
     fn sign(&self, msg: &[u8], _principal: &Principal) -> Result<Signature, String> {
         let signature = self.key_pair.sign(msg.as_ref());
@@ -60,6 +65,25 @@ impl Identity for BasicIdentity {
         Ok(Signature {
             signature: signature.as_ref().to_vec(),
             public_key: public_key_bytes.as_ref().to_vec(),
+            der_encoded_public_key: self.der_encoded_public_key.clone()
         })
     }
+}
+
+fn der_encode_public_key(public_key: Vec<u8>) -> Result<Vec<u8>, ASN1EncodeErr> {
+    // see Section 4 "SubjectPublicKeyInfo" in https://tools.ietf.org/html/rfc8410
+
+    let id_ed25519 = OID::new(vec![
+        BigUint::from(1u32),
+        BigUint::from(3u32),
+        BigUint::from(101u32),
+        BigUint::from(112u32),
+    ]);
+    let algorithm = Sequence(0, vec![ObjectIdentifier(0, id_ed25519)]);
+    let subject_public_key = BitString(0, public_key.len() * 8, public_key);
+    let subject_public_key_info = Sequence(0, vec![algorithm, subject_public_key]);
+    let x = to_der(&subject_public_key_info);
+    eprintln!("key bytes: {:?}", &x.clone().unwrap().iter()
+        .map(|x|format!("{:02X}",x)).collect::<Vec<String>>());
+    x
 }
